add and refactor root tests
