+++ cmd/secret/cmd/root_test.go
@@ -50,36 +50,6 @@ func TestRoot_Server(t *testing.T) {
 			require.EqualValues(t, http.StatusNoContent, resp.StatusCode)
 			require.NoError(t, resp.Body.Close())
 		})
-		t.Run("expect redis set method success", func(t *testing.T) {
-			ctx, cancel := context.WithTimeout(context.Background(), 15*time.Second)
-			defer cancel()
-
-			port, err := GetFreePort()
-			require.NoError(t, err)
-			r := New()
-			r.cmd.SetArgs([]string{"server", "--port", strconv.Itoa(port), "--redis-url", redisURL})
-			go func() {
-				err := r.Execute(ctx)
-				if err != nil {
-					fmt.Println(err)
-				}
-			}()
-			time.Sleep(2 * time.Second)
-			defer func() {
-				require.NoError(t, os.Remove("file.txt"))
-			}()
-
-			client := http.Client{Timeout: time.Second}
-			body := bytes.NewBufferString(`{"getter":"key-value","method":"remote","value":"test-value-1"}`)
-			req := httptest.NewRequest(http.MethodPost, "http://localhost:"+strconv.Itoa(port), body)
-			req.Header.Set(api.ParamCipherKey, expectedSipherKey)
-			req.RequestURI = ""
-
-			resp, err := client.Do(req)
-			require.NoError(t, err)
-			require.EqualValues(t, http.StatusNoContent, resp.StatusCode)
-			require.NoError(t, resp.Body.Close())
-		})
 		t.Run("expect url not found error", func(t *testing.T) {
 			ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
 			defer cancel()
@@ -95,36 +65,6 @@ func TestRoot_Server(t *testing.T) {
 			resp, err := client.Do(req)
 			require.NoError(t, err)
 			require.EqualValues(t, http.StatusNotFound, resp.StatusCode)
-			require.NoError(t, resp.Body.Close())
-		})
-		t.Run("expect postgres set method success", func(t *testing.T) {
-			ctx, cancel := context.WithTimeout(context.Background(), 15*time.Second)
-			defer cancel()
-			defer migrateDown()
-			port, err := GetFreePort()
-			require.NoError(t, err)
-			r := New()
-			r.cmd.SetArgs([]string{"server", "--port", strconv.Itoa(port), "--postgres-url", postgresURL, "--migration", "file://../../../"})
-			go func() {
-				err := r.Execute(ctx)
-				if err != nil {
-					fmt.Println(err)
-				}
-			}()
-			time.Sleep(2 * time.Second)
-			defer func() {
-				require.NoError(t, os.Remove("file.txt"))
-			}()
-
-			client := http.Client{Timeout: time.Second}
-			body := bytes.NewBufferString(`{"getter":"key-value","method":"remote","value":"test-value"}`)
-			req := httptest.NewRequest(http.MethodPost, "http://localhost:"+strconv.Itoa(port), body)
-			req.Header.Set(api.ParamCipherKey, expectedSipherKey)
-			req.RequestURI = ""
-
-			resp, err := client.Do(req)
-			require.NoError(t, err)
-			require.EqualValues(t, http.StatusNoContent, resp.StatusCode)
 			require.NoError(t, resp.Body.Close())
 		})
 	})
@@ -155,183 +95,6 @@ func TestRoot_Server(t *testing.T) {
 			query := req.URL.Query()
 			query.Set(api.ParamGetterKey, key)
 			query.Set(api.ParamMethodKey, "local")
-			req.URL.RawQuery = query.Encode()
-
-			resp, err = client.Do(req)
-			require.NoError(t, err)
-			_, err = ioutil.ReadAll(resp.Body)
-			require.NoError(t, err)
-			require.EqualValues(t, http.StatusOK, resp.StatusCode)
-			require.NoError(t, resp.Body.Close())
-		})
-		t.Run("expect bad request status if set local method and try get by remote method", func(t *testing.T) {
-			ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
-			defer cancel()
-
-			port, err := GetFreePort()
-			require.NoError(t, err)
-			r := New()
-			r.cmd.SetArgs([]string{"server", "--port", strconv.Itoa(port), "--redis-url", redisURL})
-			go func() {
-				err := r.Execute(ctx)
-				if err != nil {
-					fmt.Println(err)
-				}
-			}()
-			time.Sleep(2 * time.Second)
-
-			defer func() {
-				require.NoError(t, os.Remove("file.txt"))
-			}()
-
-			client := http.Client{Timeout: time.Second}
-			body := bytes.NewBufferString(`{"getter":"key-value","method":"local","value":"test-value-2"}`)
-			req := httptest.NewRequest(http.MethodPost, "http://localhost:"+strconv.Itoa(port), body)
-			req.Header.Set(api.ParamCipherKey, expectedSipherKey)
-			req.RequestURI = ""
-			resp, err := client.Do(req)
-			require.NoError(t, err)
-			require.NoError(t, resp.Body.Close())
-
-			req = httptest.NewRequest(http.MethodGet, "http://localhost:"+strconv.Itoa(port), nil)
-			req.RequestURI = ""
-			req.Header.Set(api.ParamCipherKey, expectedSipherKey)
-			query := req.URL.Query()
-			query.Set(api.ParamGetterKey, key)
-			query.Set(api.ParamMethodKey, "remote")
-			req.URL.RawQuery = query.Encode()
-
-			resp, err = client.Do(req)
-			require.NoError(t, err)
-			data, err := ioutil.ReadAll(resp.Body)
-			require.NoError(t, err)
-			require.Contains(t, string(data), "test-value-1")
-			require.NoError(t, resp.Body.Close())
-		})
-		t.Run("expect bad request status if set remote method and try get by local method", func(t *testing.T) {
-			ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
-			defer cancel()
-
-			port, err := GetFreePort()
-			require.NoError(t, err)
-			r := New()
-			r.cmd.SetArgs([]string{"server", "--port", strconv.Itoa(port), "--redis-url", redisURL})
-			go func() {
-				err := r.Execute(ctx)
-				if err != nil {
-					fmt.Println(err)
-				}
-			}()
-			time.Sleep(2 * time.Second)
-			defer func() {
-				require.NoError(t, os.Remove("file.txt"))
-			}()
-
-			client := http.Client{Timeout: time.Second}
-			body := bytes.NewBufferString(`{"getter":"key-value","method":"remote","value":"test-value-1"}`)
-			req := httptest.NewRequest(http.MethodPost, "http://localhost:"+strconv.Itoa(port), body)
-			req.Header.Set(api.ParamCipherKey, expectedSipherKey)
-			req.RequestURI = ""
-			resp, err := client.Do(req)
-			require.NoError(t, err)
-			require.NoError(t, resp.Body.Close())
-
-			req = httptest.NewRequest(http.MethodGet, "http://localhost:"+strconv.Itoa(port), nil)
-			req.RequestURI = ""
-			req.Header.Set(api.ParamCipherKey, expectedSipherKey)
-			query := req.URL.Query()
-			query.Set(api.ParamGetterKey, key)
-			query.Set(api.ParamMethodKey, "local")
-			req.URL.RawQuery = query.Encode()
-
-			resp, err = client.Do(req)
-			require.NoError(t, err)
-			data, err := ioutil.ReadAll(resp.Body)
-			require.NoError(t, err)
-			require.Contains(t, string(data), "cannot get data by key")
-			require.NoError(t, resp.Body.Close())
-		})
-		t.Run("expect redis get method success", func(t *testing.T) {
-			ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
-			defer cancel()
-
-			port, err := GetFreePort()
-			require.NoError(t, err)
-			r := New()
-			r.cmd.SetArgs([]string{"server", "--port", strconv.Itoa(port), "--redis-url", redisURL})
-			go func() {
-				err := r.Execute(ctx)
-				if err != nil {
-					fmt.Println(err)
-				}
-			}()
-			time.Sleep(2 * time.Second)
-			defer func() {
-				require.NoError(t, os.Remove("file.txt"))
-			}()
-
-			client := http.Client{Timeout: time.Second}
-			body := bytes.NewBufferString(`{"getter":"key-value","method":"remote","value":"test-value-1"}`)
-			req := httptest.NewRequest(http.MethodPost, "http://localhost:"+strconv.Itoa(port), body)
-			req.Header.Set(api.ParamCipherKey, expectedSipherKey)
-			req.RequestURI = ""
-
-			resp, err := client.Do(req)
-			require.NoError(t, err)
-			require.EqualValues(t, http.StatusNoContent, resp.StatusCode)
-			require.NoError(t, resp.Body.Close())
-
-			req = httptest.NewRequest(http.MethodGet, "http://localhost:"+strconv.Itoa(port), nil)
-			req.RequestURI = ""
-			req.Header.Set(api.ParamCipherKey, expectedSipherKey)
-			query := req.URL.Query()
-			query.Set(api.ParamGetterKey, key)
-			query.Set(api.ParamMethodKey, "remote")
-			req.URL.RawQuery = query.Encode()
-
-			resp, err = client.Do(req)
-			require.NoError(t, err)
-			_, err = ioutil.ReadAll(resp.Body)
-			require.NoError(t, err)
-			require.EqualValues(t, http.StatusOK, resp.StatusCode)
-			require.NoError(t, resp.Body.Close())
-		})
-		t.Run("expect postgres get method success", func(t *testing.T) {
-			ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-			defer cancel()
-			defer migrateDown()
-			port, err := GetFreePort()
-			require.NoError(t, err)
-			r := New()
-			r.cmd.SetArgs([]string{"server", "--port", strconv.Itoa(port), "--postgres-url", postgresURL, "--migration", "file://../../../"})
-			go func() {
-				err := r.Execute(ctx)
-				if err != nil {
-					fmt.Println(err)
-				}
-			}()
-			time.Sleep(2 * time.Second)
-			defer func() {
-				require.NoError(t, os.Remove("file.txt"))
-			}()
-
-			client := http.Client{Timeout: time.Second}
-			body := bytes.NewBufferString(`{"getter":"key-value","method":"remote","value":"test-value-1"}`)
-			req := httptest.NewRequest(http.MethodPost, "http://localhost:"+strconv.Itoa(port), body)
-			req.Header.Set(api.ParamCipherKey, expectedSipherKey)
-			req.RequestURI = ""
-
-			resp, err := client.Do(req)
-			require.NoError(t, err)
-			require.EqualValues(t, http.StatusNoContent, resp.StatusCode)
-			require.NoError(t, resp.Body.Close())
-
-			req = httptest.NewRequest(http.MethodGet, "http://localhost:"+strconv.Itoa(port), nil)
-			req.RequestURI = ""
-			req.Header.Set(api.ParamCipherKey, expectedSipherKey)
-			query := req.URL.Query()
-			query.Set(api.ParamGetterKey, key)
-			query.Set(api.ParamMethodKey, "remote")
 			req.URL.RawQuery = query.Encode()
 
 			resp, err = client.Do(req)
